# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Retro Gamer Blog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: pinned
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (admin/owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like/unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{id}/view:
    post:
      operationId: viewPost
      tags: [posts]
      summary: Increment post view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN, OWNER]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [USER, ADMIN, OWNER]
    Post:
      type: object
      required: [id, title, slug, content, authorId, author, createdAt, updatedAt, viewCount, likeCount, commentCount, isPinned, isFeatured]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        categoryId:
          type: ["integer", "null"]
        category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
        tags:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
        isFeatured:
          type: boolean
        viewCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        likedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostsPage:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PostInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        categoryId:
          type: integer
        tags:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
        isFeatured:
          type: boolean
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        categoryId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
        isFeatured:
          type: boolean
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        postCount:
          type: integer
...