# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Retro Gamer Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  title?: string;
  content?: string;
  excerpt?: string;
  coverImageUrl?: string;
  /** @nullable */
  categoryId?: number | null;
  tags?: string[];
  isPinned?: boolean;
  isFeatured?: boolean;
}